python/requirements.txt
grpcio>=1.48.0
grpcio-tools>=1.48.0
protobuf>=3.20.0
numpy>=1.21.0
